TSDumper: Dump out the constructors for all global variables
This commit adds the ability to dump out all the constructors on global variables, as well as generally renames "classes" to "global variables" since this is a bit more appropriate. We have also introduced a TypeInfo interface which encapsulates information such as union variants for each parameter. We have some type AST nodes that will currently output "unhandled" or "unhandled_literal" as a type name, but this can be a placeholder until we begin parsing interfaces.