TSDumper: Adding method and type parameter parsing
This commit adds support for method declarations, including type parameters for those methods. Although type parameters also apply to interfaces, this does not yet cover that. Now that we are parsing method data, we have plenty of "unhandled" and "unhandled_literal" cases in extractSingleTypeInfo that will need to be addressed.