TSDumper: Support for interface type parameters
This commit adds support for parsing interface type parameters, as well as fixing an issue with detecting constraints that was caused by an overly aggressive check for anyConstraintsAreNotSimple. Prior to this, anything other than TypeReferenceNode's were considered "not simple". Now, we really only consider a constraint not to be simple if it is a TypeOperatorNode.