TSDumper: Support for function parameter types
In addition to adding support for function parameter types, this commit also returns the entire `ExtractTypeParametersResult` rather than attempting to filter it out. It will still be able to output whether or not a type parameter is generated that is deemed overly complex. However, we will likely need to filter this out in a post-processing step instead as it will be much less complex to do after the entire output has been generated.