CodeGenerator: Initial commit
This is the initial commit for the code generator, that takes in two paths as arguments. The first path is the path to the `TSDumper` output directory. The second path is the path to where the tool should output the code that is generated. Right now, all we do is validate the paths that are inputted, and verify that a file exists for `lib.dom.d.json` inside the `TSDumper` output directory.

You may run the tool with the following command:
```
dotnet run --project CodeGenerator.csproj /path/to/TSDumper/output /path/to/CodeGenerator/output
```