TSDumper: Fixing array type extraction
The previous hack for detecting array types was broken, and was returning a `SingleTypeInfo` with a name of `[object Object][]`. It also didn't assume any support for multi-dimensional array types. This commit introduces a property inside `TypeInfo` called `array`, which similar to `parenthesized` will encapsulate an array element type definition. For multi-dimensional arrays, it would look sort of like this:

```
'type': {
  ...
  'array': {
    ...
    'array': {
      'single': {
        'name': 'string',
        ...
      },
      ...
    }
  }
}
```