TSDumper+CodeGenerator: Interface generic type arguments
This patch adds supports for the interface generic type arguments. Currently, since we do not support default values for generics, we are generating a compile error when any of these are used. Since C# doesn't have default generic types, we will have to consider overloading the interface definitions (or possibly an alternative if there is a better ideas!