TSDumper+CodeGenerator: Refactor global variables to share inner type definitions with interface body
It turns out the entire time I haven't been realizing the the following:
```
declare var Foo = {
  hello: string;
}
```

Is actually the same as:
```
declare var Foo = Bar;

interface Bar {
  hello: string;
}
```

The nice thing about this is that it means the entire time I've been thinking of global variables as something different or unique to interfaces. In reality, the top syntax is really just inlining an interface definition.

This means that we can assume that interfaces can declare prototypes and constructors. It also means that it is probably better to change the way that we think about prototype discovery for interfaces.

Although this commit doesn't address the specifics of that last detail, we haven't actually written any code to generate prototypes so we kind of dodged a bullet and caught this change before it became a bigger issue. This also explains why previously I was unable to find the prototype for `Date`!