CodeGenerator: Stub out globals and prototypes
This generates a ton of compile errors since we aren't implementing interfaces, but it is a start!

We introduced some hacky code to prevent duplicating globals (for example `window.location` is defined in the `Window` interface and globally for convenience). In the future, it'd be nice to avoid hardcoding `Window` as the globalThis context, but that should be fine for now.