Interop: Bundle a JS script to act as a bridge
This commit bundles a script that bridges the library with Javascript and provides helpers that enable the interop to work properly. The bridge is set on the `Window` object as `BlazorJavascript`, but each of the methods are also exposed individually (for example: `window.__blazorJavascript_obtainPrototype` is an alias for `window.BlazorJavascript.obtainPrototype`.)

Currently, the script is hardcoding a list of prototypes into the `obtainPrototype`, which should really be generated by CodeGenerator. This should be fine for now to get us started.