Interop: Basic implementations for primitive built-ins
Adds support and implementations for the basic primitives, as well as the most important interop methods for them (number, string, boolean, object, and function).

Most of this is built on top of extension methods for the JS runtime (IJSInProcessRuntime). Using the runtime, as well as the methods exposed by the javascript bridge, we are able to determine the prototype of object that are returned from the JS interop and construct the appropriate C# type. Similarly, using the runtime and the bridge we are able to construct javascript objects using existing C# values.