CodeGenerator+Interop: Remove the camel casing when sanitizing names
In order to prepare for global variable support, we really need to ensure that we aren't camel casing method names and properties, etc. The reality is that we have a lame tradeoff to make: We either need to come up with a convention for identifying globals that will feel unnatural to those familiar with Javascript, or we need to just accept that Javascript has lots of instances where the casing of method names/property names/globals tend to collide.