PagesGenerator: Tool to generate our Github pages in our pipeline
This tool will actually generate the files that will be uploaded as Github Pages in the build stage for our main branch. It will generate a couple simple pages:
- Index page
- Stats page

The stats page will show a bunch of graphs reflecting the progress that we are making in supporting the various use cases for our binding generation. For example, every commit we run the CodeGenerator and keep track of how many interfaces are being generated. Over time, we should be generating more interfaces, since we are going to be skipping less of them as we tackle stuff such as union types, etc.