CodeGenerator: Rewrite unions to get rid of null
We still can't handle full on unions, but by ridding our unions of "null", we can actually generate a bunch more functions because the unions can be collapsed to simple types. What this means is that if you have a method like this:
```
foo(x: Dog | null): number;
```

We can rewrite the representation as:
```
foo(x: Dog): number;
```

And now, we no longer have a union in this case.