CodeGenerator+Interop: Constructor functions for arrays and basic test case
This patch ensures that we can create arrays and do some basic operations on them (ie: check their length, push items, retrieve items by items).

We still do not generate methods that use arrays as parameter or return types, but this is going to be a step in the right direction.