CodeGenerator: Decouple `IsFinalTypeTooComplexToRender` from `ProcessTypeAliases...`
Typically, we use `IsFinalTypeTooComplexToRender` to determine if the type has anything that we are not able to handle yet (ie: unions, intersections, arrays, etc). However, we had a code smell here where the first thing that the method was doing was calling `ProcessTypeAliases...`. Instead of baking this into the method call, we decided to introduced a `ProcessedTypeInfo` (which will strongly encode our intent here, and force callers of `IsFinalTypeTooComplexToRender` to pass in a result from `ProcessTypeAliases...` instead.