CodeGenerator+Interop: Typed array support and constructor/type identification support for generics
Prior to this, we were unable to construct anything with a meaningful generic type argument. For example, arrays were simply constructed as arrays of JS objects. We still support this case, but we also now support passing both `JSFunction.Invoke` and `IJSObjectExtensions.CallConstructor` the generic arguments to use for the return type.