CodeGenerator+Interop: Supporting generics and eliminating runtime type checks
This commit is a big pivot that will help us support implementing generics properly. Prior to this, we were generating a bunch of code that was able to detect the prototype of objects that were returned from Javascript, then using that information to construct specific C# types.

This was nice because it ensured that if some foreign object was returned from Javascript that we had hints about, we could still construct the appropriate type for it. But it made it really hard to support the true duck-typing nature of Javascript.

This became really obvious when we started trying to use `getElementsByClassName`, which returns an `IHTMLCollectionOf<Node>`. What we started to notice was that the prototype of the returned object was being detected as `HtmlCollection` (which, for what it is worth, is correct!). The problem was that since we start with TypeScript definitions, we really only have enough information to infer implemented interfaces from the type definitions themselves. For example, `HTMLCollection` is not explicitly implementing `IHTMLCollectionOf<Node>` in the type definition. It does, however, _implicitly_ implement this interface. TypeScript is really clever and understands this at compile time.

I tried as best as I could to emulate this behavior through brute-force in the CodeGenerator. There are several reasons why it didn't work. The first reason is that it is extremely expensive to figure out what is being implicitly implemented by any given prototype. You essentially have to loop through all interfaces and check if you happen to implement the exact set of methods, properties, accessors, and constructors for the given interface. Next, you have to repeat the exact same process for all variants of those interfaces (any with generics, there are an infinite number of possible versions of this).

Even when I tried to hack together this brute-force mechanism, I was extremely uneasy with the nature of the tight coupling that it created. If we want any hope for this package to be generalized for libraries, we really can't take this technique very seriously.

So what are we doing instead now? We are going to stop trying to infer the types of foreign objects from Javascript. It is quite simple: We either know the type that the object _should_ be *explicitly* from the type definitions, or we don't. And when we don't, there is nothing wrong with using `JSObject` and allowing the users to cast it using `UnsafeCastTo` with any type that implements `IJSObject`. Ultimately, Javascript is designed for this type of handling of objects. This will be especially helpful for user-defined objects that conform to any interface that a function claims to accept.

Lastly, we must provide an implementation of all interfaces as a result. We were already doing this for any interface that was known to have a prototype. The only difference is now we generate a class for things that do not have a prototype. At the end of the day, this works just fine.