CodeGenerator: Refactor type simplification and aliasing to a `TypeSimplifier`
This is the first of a few commits that will decouple the generation of code from the processing of the input data. The main purpose of this commit specifically is to ensure that all `TypeInfo` processing happens before any actual code generation happens.

Prior to this, we were processing types on-demand, and this was leading to un-necessary confusion which led to additional complexity. On a side note, we now simplify types more aggressively (and accurately!) This should bump our stats a little bit.