TSDumper: Detect if a parameter is using the `...` (aka dot dot dot token) modifier
TypeScript function parameters can be annotated with `...` when they are the last parameter for a signature. This allows them to be interpreted as an array of values, where you can effectively pass any number of parameters to the argument and they will be interpreted as an array.

Let's track this modifier so we can take advantage of it when we generate code.