CodeGenerator+Interop: Add support for `...` token and ditch built-ins in favor of extension methods
Methods that take a parameter which is considered a `...` token parameter are now supported (which generates a `param T[] name` construct in C#).

We are also generally ditching built-ins in favor of extension methods, since it will mean that we have to worry even less about falling behind in support for browser APIs.