CodeGenerator+Interop: Add support for passing .NET functions to Javascript
This commit adds two main extension methods to `IJSInProcessRuntime` that allow you to create instances of `IFunction` from `Action<...>` and `Func<....>` reference types. This allows you to effectively pass in .NET functions as callbacks to Javascript land.

If anyone feels like it, we have quite a bit of code duplication in CodeGenerator now that could be cleaned up :)