CodeGenerator: Implementing a slower but safer version of `GroupBy`
Under the hood, GroupBy utilizes a HashMap/GetHashCode to group items quickly. This is great for performance, but it is prone to mistakes. Since the grouping that we are doing in the code generator doesn't need to be extremely fast but does need to be accurate 100% of the time, let's introduce a version that is much slower but safer.