TSVersionChecker+Workflows: Implement Typescript version checker
This commit adds a script executed in the CI/CD environment that bumps the version of the Interop project if there is a newer version of Typescript. A major Typescript version change will bump the major version of the Interop, and a minor OR patch version change will bump the minor version of the Interop project. The script sets the GitHub Actions specific output variables, and does not actually commit or save the bumped version in any way. The bumped version is potentially used further on in the pipeline to create a pull request to actually bump the Interop project's version.