Workflow: Use the service PAT directly instead of using `gh auth login`
When attempting to use the Github CLI in the "Update Typescript" workflow, we are not allowed to login with `gh auth login` without elevating the scope of the PAT beyond our comfort level. However, we can skip the login all together and instead set the `GH_TOKEN` environment variable prior to executing specific CLI commands. This should enable us to keep using a very restrictive token while taking advantage of the CLI.